hook up slideshow page to the backend
