+++ frontend/src/components/Slideshow.js
@@ -1,57 +1,43 @@
-import { useEffect, useState } from 'react';
-import './Slideshow.css';
-import Carousel from 'react-material-ui-carousel';
-import { Card, CardHeader } from '@mui/material';
+import { useEffect, useState } from "react";
+import "./Slideshow.css";
+import Carousel from "react-material-ui-carousel";
+import { Card, CardContent, CardHeader, Typography } from "@mui/material";
+import { fetchSlides } from "../utils/slide_utils";
+import dayjs from "dayjs";
 
-export default function Slideshow({slideStyles, slideshowProps}) {
-     const [sildeshowData, setSlideshowData] = useState([]);
-     useEffect(() => {
-        //TODO: get slideshow data from db
+export default function Slideshow({ slideStyles, slideshowProps }) {
+  const [sildeshowData, setSlideshowData] = useState([]);
+  useEffect(() => {
+    const today = dayjs().startOf("day");
+    fetchSlides(today, today.add(7, "day")).then((res) =>
+      setSlideshowData(res)
+    );
+  }, []);
 
-        setSlideshowData([
-            {
-                id: 0,
-                type: "event",
-                title: "title1",
-                description: "description1",
-                bg: "lightblue"
-            },
-            {
-                id: 1,
-                type: "menu",
-                title: "title2",
-                description: "description2",
-                bg: "lightcoral"
-            },
-            {
-                id: 2,
-                type: "event",
-                title: "title3",
-                description: "description3",
-                bg: "lightgreen"
-            },
-            {
-                id: 3,
-                type: "menu",
-                title: "title4",
-                description: "description4",
-                bg: "lightsalmon"
-            },
-        ]);
-     },[]);
-
-    return (
-        <>
-            {/* https://github.com/Learus/react-material-ui-carousel#props */}
-            <Carousel {...slideshowProps}>
-                {sildeshowData.map((slide) => {
-                    return (
-                        <Card className='slide' sx={{height:'100%', boxShadow:'0', ...slideStyles, backgroundColor: slide.bg}}>
-                            <CardHeader title={slide.title} subheader={slide.description} />
-                        </Card>
-                    )
-                })}
-            </Carousel>
-        </>
-    )
+  return (
+    <>
+      {/* https://github.com/Learus/react-material-ui-carousel#props */}
+      <Carousel {...slideshowProps}>
+        {sildeshowData.map((slide) => {
+          return (
+            <Card
+              className="slide"
+              sx={{
+                height: "100%",
+                boxShadow: "0",
+                ...slideStyles,
+              }}
+            >
+              <CardHeader title={slide.title} subheader={slide.subheader} />
+              <CardContent>
+                <Typography sx={{ fontSize: "0.6em", paddingTop: "16px" }}>
+                  {slide.body}
+                </Typography>
+              </CardContent>
+            </Card>
+          );
+        })}
+      </Carousel>
+    </>
+  );
 }
